requirements.txt
aniso8601==7.0.0
beautifulsoup4==4.6.0
boto3==1.11.3
botocore==1.14.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.8
django-allauth==0.41.0
django-contrib-comments==1.9.2
django-debug-toolbar==2.1
django-dramatiq==0.8.0
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-import-export==2.0.1
django-js-asset==1.2.2
django-modelcluster==5.0.1
django-mptt==0.10.0
django-polymorphic==2.1.2
django-storages==1.8
django-taggit==1.2.0
django-treebeard==4.3.1
django-wkhtmltopdf==3.2.0
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
dramatiq==1.7.0
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.2.1
graphql-relay==2.0.1
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
jmespath==0.9.4
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.3
Pillow==6.2.2
prometheus-client==0.7.1
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-pdf==0.37
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
s3transfer==0.3.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7.1
wagtailfontawesome==1.1.4
webencodings==0.5.1
Willow==1.3
xlrd==1.2.0
xlwt==1.3.0
